requirements.txt
google-cloud-storage
flask

